adding __call() magic method
Added the __call magic method so properties can be set through chaining methods. Also made a minor change to
the offsetGet method to only append the prefix to the value if its set. I do this because using __call causes
the property to be set to "true" and appending an empty prefix to that causes it to become the string value "1".
This way the true value is returned.