adding base grammar classes
The base grammar classes will provide generic support for SQL grammar files. The actual grammar will be
defined in the Query and Schema directories. Any methods that are generic and can be used in both areas
should be defined in this base Grammar area. This is a totally abstract section as the Grammar must be
defined in the Query or Schema.