updating namespaces and adding exceptions
Added an exception for the container for missing sections. Also moved the Config\Registry exceptions into the
Config\Registry namespace so they're easier to handle and identify later. I've also updated the tests to make
the exceptions thrown easier to detect. All this comes with minor updates to api docs.