huge rewrite of the config registry
I did not like the direction I took with the original version of the registry. This is a complete
overhaul. I moved away from using the container and wrote it to match Python's ConfigParser module.

I also added a new helper trait called Env. It will make it easier to integrate environments into
different parts of SiTech or even the application itself. I added also  another base exception
class to the helper that the config registry uses for unexpected values. I also added more
exceptions to the config registry itself to help aid in error messages when sections or options
are not found in the configuration.