updating test to include 100% coverage for registry
Fixed some minor bugs in the Registry class. One changed how addSection() and section() both work. The
section() method now has an optional force argument. If it is set to true, the "correct" section name
will be returned no matter if it exists or not. This changed the addSection() method to rely on
section() instead of hasSection() and also made it work properly for adding sections when the env is set.