major documentation update. also changed how a few methods work.
I fixed the way get() works with the default value so the documentation will actually match up
with the function declaration. This is also optimized a bit better too by simply checking the
number of arguments passed to the method. This also allows for null to actually be passed as a
default value.

I updated the getFloat/getInteger/getBoolean methods to actually call $this->get() instead of
using call_user_func_array() to call them. This preserves the stack when debugging. I also removed
some of the extra arguments that shouldn't be used in those helper methods.

The last change was how the interpolation works. Now it truly checks if interpolation should be
done based on the class variable and the variable passed into interpolate(). The unit tests
need to be updated to reflect this change. I should also break the tests out to test interpolation
seperately to ensure all scenarios are covered.